Add unit tests to the project.
Adding tests is the first thing to do before refactoring. Use JUnit as the testing framework. The tests coverage is 100% on the statement method. Make the tests launchable by gradle and explain how in the readme. Add some IntelliJ files related to sync with Gradle to gitignore. Now ready to refactor!